Intercede metadata enhancements
Gathering all metadata up front on initialisation can cause long delays (typically ~25 seconds depending on network). These changes address this problem by allowing metadata, for a particular aaguid, to be obtained on demand. Once obtained they are cached in the usual way but the time delay is reduced considerably.
They also contain an improved file system based repository that can work on demand by reading from the file system to distributed memory cache. This allows systems that do not want internet connectivity to be able to use metadata downloaded to the file system together with an optional table of contents file.
We also include a disable CRL check option to allow CRL checking to be bypassed and an optional TOC root certificate to be provided should the baked-in one become obsolete.